アニメーションを別ファイルに切り出した
